requirements.txt
ipykernel
python-dotenv
langchain-core
langchain-community
langgraph
unstructured
python-magic
python-pptx
langchain-qdrant

#AI assistant devpendencies

# Image processing 
pillow
transformers
torch
pytesseract
sentence-transformers

